Moved all source files into a new directory src.
This makes it a lot easier to run the instrumentation for the coverage as
one can specify /src as the source directory. It also makes sense to have
the source code in a sub-directory, I think?

The data directory was not moved into ./src. References to this folder has
been updated (main.js). Other than that all paths should be either relative
or for node_modules, both of which should still work the same.